requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
langchain-text-splitters
openai
python-dotenv
fastapi
langserve
sse-starlette
langchain-ollama
uvicorn